+++ src/views/days-list.js
@@ -1,4 +1,4 @@
-import createElement from "../services/utils/createElement";
+import createElement from "../utils/createElement";
 
 const createDaysListTemplate = () => `<ul class="trip-days"></ul>`;
